Separate out code, more tests
