+++ packages/eas-cli/src/api.ts
@@ -1,6 +1,6 @@
-import { JSONValue } from '@expo/json-file';
 import got, { HTTPError, NormalizedOptions, RequestError } from 'got';
 
+import ApiV2Error from './ApiV2Error';
 import { getAccessToken, getSessionSecret } from './user/sessionStorage';
 
 export const apiClient = got.extend({
@@ -37,31 +37,6 @@ export const apiClient = got.extend({
     ],
   },
 });
-
-export class ApiV2Error extends RequestError {
-  readonly name = 'ApiV2Error';
-  readonly expoApiV2ErrorCode: string;
-  readonly expoApiV2ErrorDetails?: JSONValue;
-  readonly expoApiV2ErrorServerStack?: string;
-  readonly expoApiV2ErrorMetadata?: object;
-
-  constructor(
-    originalError: HTTPError,
-    response: {
-      message: string;
-      code: string;
-      stack?: string;
-      details?: JSONValue;
-      metadata?: object;
-    }
-  ) {
-    super(response.message, originalError, originalError.request);
-    this.expoApiV2ErrorCode = response.code;
-    this.expoApiV2ErrorDetails = response.details;
-    this.expoApiV2ErrorServerStack = response.stack;
-    this.expoApiV2ErrorMetadata = response.metadata;
-  }
-}
 
 export function getExpoApiBaseUrl(): string {
   if (process.env.EXPO_STAGING) {
