+++ packages/eas-cli/src/user/__tests__/actions-test.ts
@@ -1,25 +1,15 @@
-import { apiClient } from '../../api';
-import { promptAsync, selectAsync } from '../../prompts';
+import ApiV2Error from '../../ApiV2Error';
+import { promptAsync } from '../../prompts';
 import { Actor, loginAsync } from '../User';
-import {
-  UserSecondFactorDeviceMethod,
-  ensureActorHasUsername,
-  getActorDisplayName,
-  retryUsernamePasswordAuthWithOTPAsync,
-} from '../actions';
+import { ensureActorHasUsername, getActorDisplayName, showLoginPromptAsync } from '../actions';
+import { UserSecondFactorDeviceMethod, retryUsernamePasswordAuthWithOTPAsync } from '../otp';
 
 jest.mock('../../prompts');
 jest.mock('../../api');
+jest.mock('../otp');
 jest.mock('../User', () => ({
   loginAsync: jest.fn(),
 }));
-
-const logFn = jest.fn();
-const mockLog = {
-  __esModule: true, // this property makes it work
-  default: logFn,
-};
-jest.mock('../../log', () => mockLog);
 
 beforeEach(() => {
   (promptAsync as jest.Mock).mockReset();
@@ -27,14 +17,7 @@ beforeEach(() => {
     throw new Error('Should not be called');
   });
 
-  (selectAsync as jest.Mock).mockReset();
-  (selectAsync as jest.Mock).mockImplementation(() => {
-    throw new Error('Should not be called');
-  });
-
   (loginAsync as jest.Mock).mockReset();
-
-  logFn.mockReset();
 });
 
 const userStub: Actor = {
@@ -79,231 +62,50 @@ describe('ensureActorHasUsername', () => {
   });
 });
 
-describe(retryUsernamePasswordAuthWithOTPAsync, () => {
-  it('shows SMS OTP prompt when SMS is primary and code was automatically sent', async () => {
-    (promptAsync as any)
-      .mockImplementationOnce(() => ({ otp: 'hello' }))
+describe(showLoginPromptAsync, () => {
+  it('prompts for OTP when 2FA is enabled', async () => {
+    (promptAsync as jest.Mock)
+      .mockImplementationOnce(() => ({ username: 'hello', password: 'world' }))
+      .mockImplementationOnce(() => ({ otp: '123456' }))
       .mockImplementation(() => {
         throw new Error("shouldn't happen");
       });
-
-    await retryUsernamePasswordAuthWithOTPAsync('blah', 'blah', {
-      secondFactorDevices: [
-        {
-          id: 'p0',
-          is_primary: true,
-          method: UserSecondFactorDeviceMethod.SMS,
-          sms_phone_number: 'testphone',
-        },
-      ],
-      smsAutomaticallySent: true,
-    });
-
-    expect(logFn).toHaveBeenCalledWith(
-      'One-time password was sent to the phone number ending in testphone.'
-    );
-
-    expect(loginAsync as jest.Mock).toHaveBeenCalledTimes(1);
-  });
-
-  it('shows authenticator OTP prompt when authenticator is primary', async () => {
-    (promptAsync as any)
-      .mockImplementationOnce(() => ({ otp: 'hello' }))
-      .mockImplementation(() => {
-        throw new Error("shouldn't happen");
-      });
-
-    await retryUsernamePasswordAuthWithOTPAsync('blah', 'blah', {
-      secondFactorDevices: [
-        {
-          id: 'p0',
-          is_primary: true,
-          method: UserSecondFactorDeviceMethod.AUTHENTICATOR,
-          sms_phone_number: undefined,
-        },
-      ],
-      smsAutomaticallySent: false,
-    });
-
-    expect(logFn).toHaveBeenCalledWith('One-time password from authenticator required.');
-    expect(loginAsync as jest.Mock).toHaveBeenCalledTimes(1);
-  });
-
-  it('shows menu when user bails on primary', async () => {
-    (promptAsync as any)
-      .mockImplementationOnce(() => ({ otp: null }))
-      .mockImplementationOnce(() => ({ otp: 'hello' })) // second time it is prompted after selecting backup code
-      .mockImplementation(() => {
-        throw new Error("shouldn't happen");
-      });
-
-    (selectAsync as any)
-      .mockImplementationOnce(() => -1)
-      .mockImplementation(() => {
-        throw new Error("shouldn't happen");
-      });
-
-    await retryUsernamePasswordAuthWithOTPAsync('blah', 'blah', {
-      secondFactorDevices: [
-        {
-          id: 'p0',
-          is_primary: true,
-          method: UserSecondFactorDeviceMethod.AUTHENTICATOR,
-          sms_phone_number: undefined,
-        },
-        {
-          id: 'p2',
-          is_primary: false,
-          method: UserSecondFactorDeviceMethod.AUTHENTICATOR,
-          sms_phone_number: undefined,
-        },
-      ],
-      smsAutomaticallySent: false,
-    });
-
-    expect((selectAsync as any).mock.calls.length).toEqual(1);
-    expect(loginAsync as jest.Mock).toHaveBeenCalledTimes(1);
-  });
-
-  it('shows a warning when when user bails on primary and does not have any secondary set up', async () => {
-    (promptAsync as any)
-      .mockImplementationOnce(() => ({ otp: null }))
-      .mockImplementation(() => {
-        throw new Error("shouldn't happen");
-      });
-
-    await expect(
-      retryUsernamePasswordAuthWithOTPAsync('blah', 'blah', {
-        secondFactorDevices: [
-          {
-            id: 'p0',
-            is_primary: true,
-            method: UserSecondFactorDeviceMethod.AUTHENTICATOR,
-            sms_phone_number: undefined,
+    (loginAsync as jest.Mock)
+      .mockImplementationOnce(async () => {
+        throw new ApiV2Error({ code: 'testcode', request: {} } as any, {
+          message: 'An OTP is required',
+          code: 'ONE_TIME_PASSWORD_REQUIRED',
+          metadata: {
+            secondFactorDevices: [
+              {
+                id: 'p0',
+                is_primary: true,
+                method: UserSecondFactorDeviceMethod.SMS,
+                sms_phone_number: 'testphone',
+              },
+            ],
+            smsAutomaticallySent: true,
           },
-        ],
-        smsAutomaticallySent: false,
+        });
       })
-    ).rejects.toThrowError(
-      'No other second-factor devices set up. Ensure you have set up and certified a backup device.'
-    );
-  });
+      .mockImplementation(() => {});
 
-  it('prompts for authenticator OTP when user selects authenticator secondary', async () => {
-    (promptAsync as any)
-      .mockImplementationOnce(() => ({ otp: null }))
-      .mockImplementationOnce(() => ({ otp: 'hello' })) // second time it is prompted after selecting backup code
-      .mockImplementation(() => {
-        throw new Error("shouldn't happen");
-      });
+    await showLoginPromptAsync();
 
-    (selectAsync as any)
-      .mockImplementationOnce(() => -1)
-      .mockImplementation(() => {
-        throw new Error("shouldn't happen");
-      });
-
-    await retryUsernamePasswordAuthWithOTPAsync('blah', 'blah', {
-      secondFactorDevices: [
-        {
-          id: 'p0',
-          is_primary: true,
-          method: UserSecondFactorDeviceMethod.AUTHENTICATOR,
-          sms_phone_number: undefined,
-        },
-        {
-          id: 'p2',
-          is_primary: false,
-          method: UserSecondFactorDeviceMethod.AUTHENTICATOR,
-          sms_phone_number: undefined,
-        },
-      ],
-      smsAutomaticallySent: false,
-    });
-
-    expect((promptAsync as any).mock.calls.length).toBe(2); // first OTP, second OTP
-  });
-
-  it('requests SMS OTP and prompts for SMS OTP when user selects SMS secondary', async () => {
-    (promptAsync as any)
-      .mockImplementationOnce(() => ({ otp: null }))
-      .mockImplementationOnce(() => ({ otp: 'hello' })) // second time it is prompted after selecting backup code
-      .mockImplementation(() => {
-        throw new Error("shouldn't happen");
-      });
-
-    (selectAsync as any)
-      .mockImplementationOnce(() => 0)
-      .mockImplementation(() => {
-        throw new Error("shouldn't happen");
-      });
-
-    (apiClient.post as jest.Mock).mockReturnValueOnce({
-      json: () => Promise.resolve({ data: { sessionSecret: 'SESSION_SECRET' } }),
-    });
-
-    await retryUsernamePasswordAuthWithOTPAsync('blah', 'blah', {
-      secondFactorDevices: [
-        {
-          id: 'p0',
-          is_primary: true,
-          method: UserSecondFactorDeviceMethod.AUTHENTICATOR,
-          sms_phone_number: undefined,
-        },
-        {
-          id: 'p2',
-          is_primary: false,
-          method: UserSecondFactorDeviceMethod.SMS,
-          sms_phone_number: 'wat',
-        },
-      ],
-      smsAutomaticallySent: false,
-    });
-
-    expect((promptAsync as any).mock.calls.length).toBe(2); // first OTP, second OTP
-    expect((apiClient.post as jest.Mock).mock.calls[0]).toEqual([
-      'auth/send-sms-otp',
+    expect(retryUsernamePasswordAuthWithOTPAsync as jest.Mock).toHaveBeenCalledWith(
+      'hello',
+      'world',
       {
-        json: {
-          username: 'blah',
-          password: 'blah',
-          secondFactorDeviceID: 'p2',
-        },
-      },
-    ]);
-  });
-
-  it('exits when user bails on primary and backup', async () => {
-    (promptAsync as any)
-      .mockImplementationOnce(() => ({ otp: null }))
-      .mockImplementation(() => {
-        throw new Error("shouldn't happen");
-      });
-
-    (selectAsync as any)
-      .mockImplementationOnce(() => -2)
-      .mockImplementation(() => {
-        throw new Error("shouldn't happen");
-      });
-
-    await expect(
-      retryUsernamePasswordAuthWithOTPAsync('blah', 'blah', {
         secondFactorDevices: [
           {
             id: 'p0',
             is_primary: true,
-            method: UserSecondFactorDeviceMethod.AUTHENTICATOR,
-            sms_phone_number: undefined,
-          },
-          {
-            id: 'p2',
-            is_primary: false,
-            method: UserSecondFactorDeviceMethod.AUTHENTICATOR,
-            sms_phone_number: undefined,
+            method: UserSecondFactorDeviceMethod.SMS,
+            sms_phone_number: 'testphone',
           },
         ],
-        smsAutomaticallySent: false,
-      })
-    ).rejects.toThrowError('Cancelled login');
+        smsAutomaticallySent: true,
+      }
+    );
   });
 });
